requirement.txt
pytorch-cuda=11.8
tqdm
numpy
matplotlib
argparse
mpl_toolkits
scipy
seaborn
pandas